Add version, PKGPATH, arg passing to mkbuildenv.sh
The new command now accepts additional arguments after the path
which are passed into the mkbuildenv.sh script, with arg numbering
set from $1 for the first additional arg. If you want to omit the
path arg, use "--" so show you are done. Alternatively specifying
the target path as "." also works.