Added Cpyl package.
The Cpyl (C/Python library) package adds the "swig" build target
to the Clib Makefile. This builds a python module using SWIG and
named pyname. The version of Python to use, along with the header
file to use as the module's interface, are specified in config.mk.
mkbuildenv.sh has been modified to accept the project's name as a
command line argument - much more useful than the compiler choice.